pyproject.toml
[project]
name = "whatsapp-orchestrator"
version = "0.1.0"
description = "Sistema FastAPI + LangGraph para orquestração de fluxos WhatsApp"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "langgraph>=0.0.40",
    "langchain-core>=0.1.0",
    "langchain-openai>=0.0.5",
    "httpx>=0.25.0",
    "redis>=5.0.0",
    "python-dotenv>=1.0.0",
    "pinecone-client>=3.0.0",
    "gspread>=5.12.0",
    "orjson>=3.9.0",
    "structlog>=23.2.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "google-auth>=2.23.0",
    "google-auth-oauthlib>=1.1.0",
    "google-auth-httplib2>=0.2.0",
    "sentence-transformers>=2.2.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]
